Default rig now spawned on scene setup
