+++ Source/Basic-Interaction-Component/Editor/RigSetup/DefaultRigSceneSetup.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 96c28fc9213810c4bb551a8c4103b012
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
