+++ Source/XR-Interaction-Component/Source/Runtime/Configuration/XRInteractionComponentConfiguration.cs
@@ -12,5 +12,8 @@ namespace VRBuilder.XRInteraction.Configuration
 
         /// <inheritdoc/>
         public bool IsXRInteractionComponent => true;
+
+        /// <inheritdoc/>
+        public string DefaultRigPrefab => "[XR_Setup_Action_Based]";
     }
 }
